refactor: remove cycle from DocumentList
